Modify GoogleMap to use AndroidView overload with onReset lambda
This will make sure that the underlying MapView is re-used in a LazyColumn (and elsewhere) which will significantly improve scrolling performance in LazyColumn.

If this overload isn't used the MapView will be destroyed every time it leaves composition.

The MapView will still be destroyed when the parent node leaves the composition.