fix: Prevent Clustering with composable content from displaying default markers
Wait until we have both a non-null clusterManager and renderer created before configuring the cluster manager listeners. This is a small optimization that prevents setting the listeners on the default renderer when the clusterManger is first created (before the renderer is created).

Wait until the renderer is created before configuring the map for clustering. This prevents the default renderer from rendering markers.