feat: navigation compose and delegation pattern (#670)
* experimenting with navcompose

* feature: creates a delegate that can wrap a MapView.

* feat(navigation-app): Add Navigation support to Maps Compose

Adds support for the Navigation SDK to Maps Compose. It just allows the user to replace a standard MapView with a NavigationView within a GoogleMap composable.

The following changes were made:

Added a new NavigationViewDelegate class to handle the integration between NavigationView and Maps Compose.
Added a new NavigationScreen composable function to display the navigation view.
Added a new MovableMarker composable function to display a draggable marker on the map.
Updated the GoogleMap composable function to support the use of NavigationView.
Added a new NavigationApplication class to initialize the Places SDK.
Added a new ApiKeyProvider class to provide API keys for the Maps and Places SDKs.
Added a new LocationProvider class to provide location data.
Added a new PermissionChecker class to check for location permissions.
Updated the build.gradle.kts files to include the necessary dependencies.
Updated the local.defaults.properties file to include the Places API key.

* feat: Add NavigationViewDelegate

Adds NavigationViewDelegate to support NavigationView in the navigation sample app.

This change allows the navigation sample app to use NavigationView, which provides turn-by-turn navigation functionality.

The NavigationViewDelegate handles the lifecycle and rendering of the NavigationView, ensuring that it is properly integrated into the Jetpack Compose UI.

Uses a MarkerComposable to demonstrate the map really is a composable

* experimenting with navcompose

* feature: creates a delegate that can wrap a MapView.

* feat(navigation-app): Add Navigation support to Maps Compose

Adds support for the Navigation SDK to Maps Compose. It just allows the user to replace a standard MapView with a NavigationView within a GoogleMap composable.

The following changes were made:

Added a new NavigationViewDelegate class to handle the integration between NavigationView and Maps Compose.
Added a new NavigationScreen composable function to display the navigation view.
Added a new MovableMarker composable function to display a draggable marker on the map.
Updated the GoogleMap composable function to support the use of NavigationView.
Added a new NavigationApplication class to initialize the Places SDK.
Added a new ApiKeyProvider class to provide API keys for the Maps and Places SDKs.
Added a new LocationProvider class to provide location data.
Added a new PermissionChecker class to check for location permissions.
Updated the build.gradle.kts files to include the necessary dependencies.
Updated the local.defaults.properties file to include the Places API key.

* feat: Add NavigationViewDelegate

Adds NavigationViewDelegate to support NavigationView in the navigation sample app.

This change allows the navigation sample app to use NavigationView, which provides turn-by-turn navigation functionality.

The NavigationViewDelegate handles the lifecycle and rendering of the NavigationView, ensuring that it is properly integrated into the Jetpack Compose UI.

Uses a MarkerComposable to demonstrate the map really is a composable

* chore: merge agpVersion with apg in the version catalog file

* feat: added LeakCanary, code clean

* feat: added some documentation

* build(deps): Upgrade navigation library to v6.1.0

Updates the navigation library dependency from v6.0.1 to v6.1.0.

This commit also the wording of the comment in `navigation-app/build.gradle.kts` to
clarify the reason for excluding `play-services-maps`.

* feat: renamed app module to maps-app

* feat: updated lint config

---------

Co-authored-by: Enrique López Mañas <eenriquelopez@gmail.com>